Fix errors spotted by Fred
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@29662 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Foundation/NSSpellServer.h
@@ -111,7 +111,7 @@ GS_EXPORT NSString *const NSGrammarUserDescription;
 - (NSRange) spellServer: (NSSpellServer *)sender
 findMisspelledWordInString: (NSString *)stringToCheck
                   language: (NSString *)language
-                 wordCount: (NSInteger *)wordCount
+                 wordCount: (int32_t *)wordCount
                  countOnly: (BOOL)countOnly;
 
 /**
